Revert "Handle errors during the archive/unarchive process"
