+++ corehq/form_processor/submission_post.py
@@ -2,6 +2,8 @@
 from __future__ import absolute_import
 
 from __future__ import unicode_literals
+import contextlib
+import datetime
 import logging
 from collections import namedtuple
 
@@ -34,7 +36,6 @@ from corehq.form_processor.interfaces.dbaccessors import FormAccessors
 from corehq.form_processor.interfaces.processor import FormProcessorInterface
 from corehq.form_processor.parsers.form import process_xform_xml
 from corehq.form_processor.utils.metadata import scrub_meta
-from corehq.form_processor.submission_process_tracker import unfinished_submission
 from corehq.util.global_request import get_request
 from couchforms import openrosa_response
 from couchforms.const import BadRequest, DEVICE_LOG_XMLNS
@@ -516,3 +517,33 @@ def notify_submission_error(instance, exception, message):
         notify_exception(request, message, details=details)
     else:
         logging.error(message, exc_info=sys.exc_info(), extra={'details': details})
+
+
+class SubmissionProcessTracker(object):
+    def __init__(self, stub=None):
+        self.stub = stub
+
+    def submission_saved(self):
+        if self.stub:
+            self.stub.saved = True
+            self.stub.save()
+
+    def submission_fully_processed(self):
+        if self.stub:
+            self.stub.delete()
+
+
+@contextlib.contextmanager
+def unfinished_submission(instance):
+    unfinished_submission_stub = None
+    if not getattr(instance, 'deprecated_form_id', None):
+        # don't create stubs for form edits since we don't want to auto-reprocess them
+        unfinished_submission_stub = UnfinishedSubmissionStub.objects.create(
+            xform_id=instance.form_id,
+            timestamp=datetime.datetime.utcnow(),
+            saved=False,
+            domain=instance.domain,
+        )
+    tracker = SubmissionProcessTracker(unfinished_submission_stub)
+    yield tracker
+    tracker.submission_fully_processed()
